+++ examples/cycle_tokio_task.rs
@@ -11,10 +11,10 @@ fn main() -> Result<()> {
     let delay_timer = DelayTimerBuilder::default().tokio_runtime(None).build();
 
     // Develop a task that runs in an asynchronous cycle (using a custom asynchronous template).
-    delay_timer.add_task(build_task_customized_async_task())?;
+    delay_timer.add_task(build_task_customized_async_task()?)?;
 
     // Develop a task that runs in an asynchronous cycle to wake up the current thread.
-    delay_timer.add_task(build_wake_task())?;
+    delay_timer.add_task(build_wake_task()?)?;
     park();
 
     // No new tasks are accepted; running tasks are not affected.
@@ -23,20 +23,19 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn build_task_customized_async_task() -> Task {
+fn build_task_customized_async_task() -> Result<Task> {
     let mut task_builder = TaskBuilder::default();
 
     let body = generate_closure_template("'delay_timer-is-easy-to-use.'".into());
 
-    task_builder
+    Ok(task_builder
         .set_frequency_by_candy(CandyFrequency::Repeated(AuspiciousDay::Work))
         .set_task_id(5)
         .set_maximum_running_time(15)
-        .spawn(body)
-        .unwrap()
+        .spawn(body)?)
 }
 
-fn build_wake_task() -> Task {
+fn build_wake_task() -> Result<Task> {
     let mut task_builder = TaskBuilder::default();
 
     let thread: Thread = current();
@@ -46,12 +45,11 @@ fn build_wake_task() -> Task {
         create_default_delay_task_handler()
     };
 
-    task_builder
+    Ok(task_builder
         .set_frequency_by_candy(CandyFrequency::Repeated(AuspiciousDay::Wake))
         .set_task_id(7)
         .set_maximum_running_time(50)
-        .spawn(body)
-        .unwrap()
+        .spawn(body)?)
 }
 
 pub fn generate_closure_template(
@@ -61,7 +59,7 @@ pub fn generate_closure_template(
         let future_inner = async_template(get_timestamp() as i32, name.clone());
 
         let future = async move {
-            future_inner.await;
+            future_inner.await.ok();
             context.finishe_task(None).await;
         };
 
@@ -69,20 +67,21 @@ pub fn generate_closure_template(
     }
 }
 
-pub async fn async_template(id: i32, name: String) {
+pub async fn async_template(id: i32, name: String) -> Result<()> {
     let client = Client::new();
 
     //The default connector does not handle TLS.
     //Speaking to https destinations will require configuring a connector that implements TLS.
     //So use http for test.
     let url = format!("http://httpbin.org/get?id={}&name={}", id, name);
-    let uri: Uri = url.parse().unwrap();
+    let uri: Uri = url.parse()?;
 
-    let res = client.get(uri).await.unwrap();
+    let res = client.get(uri).await?;
     println!("Response: {}", res.status());
     // Concatenate the body stream into a single buffer...
-    let buf = hyper::body::to_bytes(res).await.unwrap();
+    let buf = hyper::body::to_bytes(res).await?;
     println!("body: {:?}", buf);
+    Ok(())
 }
 
 enum AuspiciousDay {