+++ src/utils/parse.rs
@@ -48,6 +48,11 @@ pub mod shell_command {
                     self
                 }
 
+                fn stderr<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self {
+                    self.stderr(cfg);
+                    self
+                }
+
                 fn spawn(&mut self) -> AnyResult<$child> {
                     Ok(self.spawn()?)
                 }
@@ -73,6 +78,9 @@ pub mod shell_command {
 
         /// Configuration for the child process's standard output (stdout) handle.
         fn stdout<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self;
+
+        /// Configuration for the child process's standard error (stderr) handle.
+        fn stderr<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self;
 
         /// Executes the command as a child process, returning a handle to it.
         fn spawn(&mut self) -> AnyResult<Child>;
@@ -222,7 +230,7 @@ pub mod shell_command {
             }
 
             let mut parts = command.trim().split_whitespace();
-            let command = parts.next().unwrap();
+            let command = parts.next().ok_or_else(|| anyhow!("Without next part"))?;
             let args = parts;
 
             // Standard input to the current process.
@@ -244,11 +252,11 @@ pub mod shell_command {
             }
 
             let mut output = Command::new(command);
-            output.args(args).stdin(stdin);
+            output.args(args).stdin(stdin).stderr(Stdio::piped());
 
             let process: Child;
-            let end_flag = if check_redirect_result.is_some() {
-                let stdout = check_redirect_result.unwrap()?;
+            let end_flag = if let Some(stdout_result) = check_redirect_result {
+                let stdout = stdout_result?;
                 process = output.stdout(stdout).spawn()?;
                 true
             } else {
@@ -264,6 +272,7 @@ pub mod shell_command {
                 //     // there should be a default file to record it.
                 //     stdout = Stdio::inherit();
                 // };
+
                 process = output.stdout(stdout).spawn()?;
                 false
             };